+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 # 0.2.0
 - Explicitly sort sidepanel images/containers by newest
+- Persist settings as YAML file in appropriate OS config directory
+- Settings will be saved periodically when using the app
+- Add `save` button to settings window